GH-206: Fix ITs for async views
ITs that need to be async to use async objects, such as repositories, fail when
exercising an async Flask view [1]. This is because pytest-asyncio runs the test
on the event loop which prevents Flask from adapting the async view to sync.

The workaround to this is to offload invoking Flask endpoints to an executor, as
they expect to be sync and thus blocking. We do this by decorating the sync
`FlaskClient` with asgiref's sync-to-async adapter to produce an async version,
as inspired by [2]. This is then used in all async ITs.

[1] https://github.com/pallets/flask/issues/4375
[2] https://stackoverflow.com/questions/69431468/can-we-use-flask-test-client-inside-async-function/75674848#75674848