GH-206: Use an async OAuth client for ATE API
Authlib doesn't provide an async OAuth client for Flask [1], so build a custom
one from the available mixins that uses HTTPX instead of Requests.

That requires reworking the workaround to support audience [2] to align with the
async implementation.

Enable async test support in Pytest and migrate from Responses to RESPX for
mocking HTTP.

[1] https://github.com/authlib/authlib/issues/818
[2] https://github.com/authlib/authlib/issues/783