GH-206: Reuse HTTPX client for persistent connection
Authlib creates a new HTTPX client for each API request [1]. This impacts
performance as a new TCP connection is created for each API request which is not
recommended [2].

To work around this, we expose the OAuth/HTTPX client from the Authlib remote
app by wrapping it in HTTP method convenience methods similar to `BaseApp`. This
allows API repositories to create a single client as a context manager and
perform multiple requests efficiently.

We test this by spying on the number of times that the stub remote app creates a
client. This should be one for a single repository method, and ideally in future
one for a single endpoint.

[1] https://github.com/authlib/authlib/issues/822
[2] https://www.python-httpx.org/advanced/clients/#why-use-a-client