GH-206: Extend ATE API timeout to cover Cloud Run startup time
Before this story we used Requests to make API requests which has no default
timeout [1]. Moving to HTTPX means that we now inherit a default timeout of 5s
[2].

In non-production environments the API is not kept idle, so it's not uncommon
for the service to wake up the API. Timeout errors have been seen at 5s, which
makes sense as the API logs show a typical startup time of 8.7s. Extend the
timeout to 10s to see if that covers it.

[1] https://requests.readthedocs.io/en/latest/user/advanced/#timeouts
[2] https://www.python-httpx.org/advanced/timeouts/