GH-189: Use resource protector to validate JWT in E2E stub API
Authlib's resource protector is the recommended way to check access tokens in a
resource server [1]. Use `JWTBearerTokenValidator` [2] to validate the JWT.

This gives us the correct response status code when no access token is provided.

Our decorator is slightly more complex as we want to obtain the issuer and
resource server identifier at runtime.

[1] https://docs.authlib.org/en/latest/flask/2/resource-server.html
[2] https://docs.authlib.org/en/latest/specs/rfc9068.html#authlib.oauth2.rfc9068.JWTBearerTokenValidator