GH-189: Provide access token in E2E API client requests
This allows the API client to invoke non-test endpoints that are protected by
OAuth.

For simplicity, we only obtain one access token for the API client to use at
construction, which is once per test run. This seems to be the recommended
approach for Requests-based Authlib OAuth clients [1], whereas the automatic
fetch and update functionality is reserved for web framework clients [2]. The
default JWT expiry for the stub authorisation server is one hour, so this
shouldn't cause a problem.

Introduce another OAuth client to the authorisation server for the E2E tests to
use. This distinguishes it from the OAuth client used by the system under test.
We can then use scopes to control which endpoints the two different clients can
invoke.

Note that we widen the API client fixture scope to package. This resolves an
issue obtaining the token endpoint dynamically from the authorisation server
Flask app. When the fixture scope defaults to function, pytest-flask pushes a
request context for the main app [3] which forces all Flask URL lookups to
resolve against that app [4], irrespective of what app instance the lookup was
invoked on. A wider scope is arguably better as it aligns with the API server
lifecycle and means a single access token can be used for the whole test run.

[1] https://docs.authlib.org/en/latest/client/oauth2.html#oauth2session-for-client-credentials
[2] https://docs.authlib.org/en/latest/client/frameworks.html#accessing-oauth-resources
[3] https://pytest-flask.readthedocs.io/en/latest/features.html#client-application-test-client
[4] https://github.com/pallets/flask/issues/1412