GH-189: Update financial revisions in the API
Determine which financial revisions are new and need to be posted to the API by
treating a missing domain id as transient. Persistent domain objects have an id
of zero, since that's what we default it to in the repository get methods,
because the API correctly doesn't expose database ids.

In this API repository implementation, the API is responsible for closing
previous revisions rather than this service's domain, as is the case with the
database repository implementation.